HAWQ-1490: Introduced text-based resolver and refactored class names
